requirements.txt
streamlit
openai
spacy
pdfplumber
python-docx
pandas
plotly
python-dotenv
google-generativeai